requirements.txt
asn1crypto==0.24.0
blinker==1.4
cffi==1.12.3
Click==7.0
cryptography==2.6.1
Faker==1.0.5
Flask==1.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.10.1
six==1.12.0
SQLAlchemy==1.3.3
text-unidecode==1.2
Werkzeug==0.15.2
WTForms==2.2.1
